+++ identity/customer_secret_key_summary.go
@@ -39,7 +39,7 @@ type CustomerSecretKeySummary struct {
 	LifecycleState CustomerSecretKeySummaryLifecycleStateEnum `mandatory:"false" json:"lifecycleState,omitempty"`
 
 	// The detailed status of INACTIVE lifecycleState.
-	InactiveStatus *int64 `mandatory:"false" json:"inactiveStatus,omitempty"`
+	InactiveStatus *int `mandatory:"false" json:"inactiveStatus,omitempty"`
 }
 
 func (model CustomerSecretKeySummary) String() string {
